+++ packages/cli/commons/src/utils/os.spec.ts
@@ -1,18 +1,17 @@
-import {fancyIt} from '@coveo/cli-commons-dev/testUtils/it';
 import {appendCmdIfWindows} from './os';
 
 describe('appendCmdIfWindows', () => {
   let originalProcess: PropertyDescriptor | undefined;
-  afterEach(() => {
+
+  beforeAll(() => {
     originalProcess = Object.getOwnPropertyDescriptor(process, 'platform');
-    jest.resetAllMocks();
-  });
-  afterAll(() => {
-    Object.defineProperty(process, 'platform', originalProcess!);
-    jest.clearAllMocks();
   });
 
-  fancyIt()('should append cmd if process.platform is win32', () => {
+  afterAll(() => {
+    Object.defineProperty(process, 'platform', originalProcess!);
+  });
+
+  it('should append cmd if process.platform is win32', () => {
     Object.defineProperty(process, 'platform', {
       writable: false,
       value: 'win32',
@@ -21,7 +20,7 @@ describe('appendCmdIfWindows', () => {
     return expect(appendCmdIfWindows`foo`).toEqual('foo.cmd');
   });
 
-  fancyIt()('should append cmd if process.platform is not win32', () => {
+  it('should append cmd if process.platform is not win32', () => {
     Object.defineProperty(process, 'platform', {
       writable: false,
       value: 'linux',