+++ packages/cli/core/src/lib/utils/cli.spec.ts
@@ -1,19 +1,18 @@
 jest.mock('@coveo/cli-commons/preconditions/trackable');
 jest.mock('@coveo/cli-commons/config/globalConfig');
+jest.mock('@coveo/cli-commons/utils/ux');
 
-import {Interfaces, CliUx} from '@oclif/core';
+import {Interfaces} from '@oclif/core';
+import {confirm} from '@coveo/cli-commons/utils/ux';
 import {fancyIt} from '@coveo/cli-commons-dev/testUtils/it';
 import globalConfig from '@coveo/cli-commons/config/globalConfig';
 import {confirmWithAnalytics} from './cli';
 
 const mockedGlobalConfig = jest.mocked(globalConfig);
 const mockedAnalyticHook = jest.fn();
-const mockedConfirm = jest.fn();
-const doMockConfirm = () => {
-  Object.defineProperty(CliUx.ux, 'confirm', {value: mockedConfirm});
-};
 
 describe('cli', () => {
+  const mockedConfirm = jest.mocked(confirm);
   beforeAll(() => {
     mockedGlobalConfig.get.mockReturnValue({
       configDir: 'the_config_dir',
@@ -21,22 +20,18 @@ describe('cli', () => {
     } as unknown as Interfaces.Config);
   });
 
-  beforeEach(() => {
-    doMockConfirm();
-  });
-
   afterEach(() => {
-    mockedConfirm.mockClear();
+    mockedConfirm.mockReset();
   });
 
   fancyIt()('should call cli.confirm with the right question', async () => {
     await confirmWithAnalytics('this is a question', 'question name');
-    expect(mockedConfirm).toHaveBeenCalledWith('this is a question');
+    expect(mockedConfirm).toHaveBeenCalledWith('this is a question', false);
   });
 
   describe('when the user confirmed', () => {
     beforeEach(() => {
-      mockedConfirm.mockReturnValue(true);
+      mockedConfirm.mockResolvedValue(true);
     });
 
     fancyIt()('should log a `confirmed` event', async () => {
@@ -52,7 +47,7 @@ describe('cli', () => {
 
   describe('when the user did not confirm', () => {
     beforeEach(() => {
-      mockedConfirm.mockReturnValue(false);
+      mockedConfirm.mockResolvedValue(false);
     });
 
     fancyIt()('should log a `cancelled` event', async () => {