+++ packages/ui/angular/src/ng-add-setup-project/files/src/environments/environment.d.ts
@@ -4,4 +4,5 @@ export declare const environment: {
   defaultTokenEndpoint: string;
   platformUrl: string;
   organizationId: string;
+  platformEnvironment: 'prod' | 'hipaa' | 'stg' | 'dev';
 };