+++ packages/ui/atomic/create-atomic-component-project/template/src/pages/index.ts
@@ -7,6 +7,9 @@ async function main() {
   await searchInterface.initialize({
     organizationId: 'barcagroupproductionkwvdy6lp',
     accessToken: 'xxfbc21fae-f1ee-45f9-8a7a-a7de938e0e05',
+    organizationEndpoints: await searchInterface.getOrganizationEndpoints(
+      'barcagroupproductionkwvdy6lp'
+    ),
   });
 
   searchInterface.executeFirstSearch();