+++ packages/ui/atomic/template/src/index.ts
@@ -5,13 +5,16 @@ async function main() {
   const searchInterface: HTMLAtomicSearchInterfaceElement =
     document.querySelector('atomic-search-interface')!;
 
-  const platformUrl = process.env.PLATFORM_URL!;
   const organizationId = process.env.ORGANIZATION_ID!;
+  const platformEnvironment = process.env.PLATFORM_ENVIRONMENT || 'prod';
   const accessToken = process.env.API_KEY!;
   await searchInterface.initialize({
     accessToken,
     organizationId,
-    platformUrl,
+    organizationEndpoints: await searchInterface.getOrganizationEndpoints(
+      organizationId,
+      platformEnvironment as 'dev' | 'stg' | 'prod' | 'hipaa'
+    ),
   });
 
   searchInterface.executeFirstSearch();