+++ packages/ui/cra-template/template/scripts/start-server.mjs
@@ -12,6 +12,7 @@ function getEnvVariables() {
       ORGANIZATION_ID: process.env.REACT_APP_ORGANIZATION_ID,
       PLATFORM_URL: process.env.REACT_APP_PLATFORM_URL,
       USER_EMAIL: process.env.REACT_APP_USER_EMAIL,
+      PLATFORM_ENVIRONMENT: process.env.REACT_APP_PLATFORM_ENVIRONMENT,
     },
     process.env
   );